requirements.txt
aiosmtpd==1.4.5
atpublic==4.1.0
attrs==23.2.0
click==8.0.1
git-filter-repo==2.38.0
numpy==1.26.4
pandas==2.2.2
pydicom==2.4.4
pynetdicom==2.0.2
python-dateutil==2.9.0.post0
pyodbc==5.1.0
pytz==2024.1
requests-mock==1.11.0
six==1.16.0
tenacity==8.3.0
toml==0.10.2
tzdata==2024.1
python-dotenv>=0.20.0
Flask-HTTPAuth>=4.0.0
marshmallow>=3.0.0
